XPulpIMG MAC and SIMD instructions

See merge request mempool/mempool!78

+++ apps/sleep_wakeup/main.c
@@ -6,7 +6,7 @@
 #include "runtime.h"
 #include "synchronization.h"
 
-volatile uint32_t atomic __attribute__((section(".l2"))) = -1;
+volatile uint32_t atomic __attribute__((section(".l2"))) = (uint32_t)-1;
 
 extern volatile uint32_t tcdm_start_address_reg;
 extern volatile uint32_t tcdm_end_address_reg;
